Use shared contexts and shared examples to minimize code duplication.
